# payment-service/src/main/java/com/example/payment/web/CreditCard.java
package com.example.payment.web;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotNull;
import javax.validation.constraints.Pattern;
import javax.validation.constraints.Size;

import org.hibernate.validator.constraints.CreditCardNumber;

public class CreditCard {
    @CreditCardNumber private String number;
    @NotBlank @NotNull private String holder;
    @NotNull @Pattern(regexp = "\\d\\d/\\d\\d") private String expiration;
    @NotBlank @NotNull @Size(min = 3, max = 3) private String ccv;
    
    public String getNumber() {
        return number;
    }
    
    public void setNumber(String number) {
        this.number = number;
    }

    public String getHolder() {
        return holder;
    }

    public void setHolder(String holder) {
        this.holder = holder;
    }

    public String getExpiration() {
        return expiration;
    }

    public void setExpiration(String expiration) {
        this.expiration = expiration;
    }

    public String getCcv() {
        return ccv;
    }

    public void setCcv(String ccv) {
        this.ccv = ccv;
    }
}